refactor(misc): scope agentic onboard to nx connect only
## Current Behavior
nx init and CNW spawn the agentic onboard bin from their own flows when an agent picks "yes" to Nx Cloud. CNW also pushes a connect URL into the workspace README and a "run nx connect" banner whenever the user picks "Maybe later" or "Skip" on the cloud prompt.

## Expected Behavior
Only nx connect drives the agentic onboard — agents typically reach connect after init/CNW anyway. nx init reverts to the original initCloud("nx-init") path; CNW reverts to openCloudSetupUrl in the post-git block. CNW only generates the connect URL / README banner when the user explicitly opts in (matches nx init). "Maybe later", "Skip", and "Never" leave the workspace clean — user (or agent) runs nx connect on their own. The onboard helper moves under nx-cloud/connect/ since connect is its only caller, and the duplicated CNW copy is deleted.

## Related Issue(s)
NXC-4401
