+++ packages/create-nx-workspace/src/utils/nx/messages.spec.ts
@@ -1,4 +1,4 @@
-import { getCompletionMessage, getSkippedCloudMessage } from './messages';
+import { getCompletionMessage } from './messages';
 import { VcsPushStatus } from '../git/git';
 
 describe('Nx Cloud Messages', () => {
@@ -303,20 +303,4 @@ describe('Nx Cloud Messages', () => {
     });
   });
 
-  describe('Skipped Cloud Messages', () => {
-    it('should return next steps message with nx connect instruction', () => {
-      const message = getSkippedCloudMessage();
-      expect(message).toMatchInlineSnapshot(`
-        {
-          "bodyLines": [
-            "Run "nx connect" to enable remote caching and speed up your CI.",
-            "",
-            "70% faster CI, 60% less compute, automatically fix broken PRs.",
-            "Learn more at https://nx.dev/nx-cloud",
-          ],
-          "title": "Next steps",
-        }
-      `);
-    });
-  });
 });
