+++ packages/nx/src/command-line/nx-cloud/connect/agentic-onboard.spec.ts
